+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Added `MeterRegistry` to `ApiClient`.
 - Added `MeterRegistry` to most methods in `BasicMeters`.
 - Changed `ClientProperties` to use `Duration` instead of `int` (milliseconds).
+- Renamed `HttpExchangeClient` to `JavaNetHttpExchangeClient` to emphasize the underlying HTTP client.
+- Added `HttpExchangeClient` interface with HTTP specific methods
 
 ---